requirements.txt
# Core dependencies
PySide6>=6.6.0

# AWS SDK
boto3>=1.34.0

# Configuration
python-dotenv>=1.0.0
